The id, generation relied on incrementing static long variable. This is not thread safe. should be ok, because id generation relies on System time and count.
git-svn-id: https://svn.apache.org/repos/asf/james/server/trunk@107445 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/james/James.java
@@ -75,7 +75,10 @@ public class James
     private String inboxRootURL;
     private UsersRepository localusers;
     private Collection serverNames;
-    private static long count;
+    // this used to be long, but increment operations on long are not
+    // thread safe. Changed to int. 'int' should be ok, because id generation
+    // is based on System time and count
+    private static int count;
     private String helloName;
     private String hostName;
     private Map mailboxes; //Not to be shared!
